+++ compiler-tests/src/test/data/box/bytecode/SwitchingProvidersGenerateTableSwitch.kt
@@ -1,4 +1,5 @@
 // ENABLE_SWITCHING_PROVIDERS: true
+// STATEMENTS_PER_INIT_FUN: 2
 
 @SingleIn(AppScope::class)
 @Inject
@@ -12,11 +13,26 @@ class Thing2
 @Inject
 class Thing3
 
+@SingleIn(AppScope::class)
+@Inject
+class Thing4
+
+@SingleIn(AppScope::class)
+@Inject
+class Thing5
+
+@SingleIn(AppScope::class)
+@Inject
+class Thing6
+
 @DependencyGraph(AppScope::class)
 interface AppGraph {
   val thing1: Thing1
   val thing2: Thing2
   val thing3: Thing3
+  val thing4: Thing4
+  val thing5: Thing5
+  val thing6: Thing6
 }
 
 fun box(): String {
@@ -24,14 +40,17 @@ fun box(): String {
   assertNotNull(graph.thing1)
   assertNotNull(graph.thing2)
   assertNotNull(graph.thing3)
+  assertNotNull(graph.thing4)
+  assertNotNull(graph.thing5)
+  assertNotNull(graph.thing6)
   return "OK"
 }
 
 // <count> <instruction>
-// Asserts that we use tableswitch and not fall-through equals checks
+// Asserts that invoke() routes directly to one helper and all switches lower to tableswitch.
 // CHECK_BYTECODE_TEXT
+// @AppGraph$Impl$SwitchingProvider.class:
 // 0 Intrinsics.areEqual
 // 0 IF_ICMPNE
-// 1 TABLESWITCH
-// 1 ISTORE 1
-// 1 GETFIELD AppGraph\$Impl\$SwitchingProvider\.id : I
+// 1 IDIV
+// 4 TABLESWITCH